+++ openpathsampling/__init__.py
@@ -55,7 +55,7 @@ from analysis.shooting_point_analysis import (
     ShootingPointAnalysis, SnapshotByCoordinateDict
 )
 
-from analysis.single_trajectory_analysis import (
+from analysis.trajectory_transition_analysis import (
     TrajectoryTransitionAnalysis,
     TrajectorySegmentContainer
 )
