+++ openpathsampling/analysis/trajectory_transition_analysis.py
