+++ openpathsampling/tests/test_trajectory_transition_analysis.py
